Enable easy local cluster setup with cpp worker
